requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.2.7
django-cloudinary-storage==0.3.0
django-filter==25.2
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.11
MarkupSafe==3.0.3
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
python-http-client==3.3.7
requests==2.32.5
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.11.0
